Enhance functionality by updating requirements, improving path management, and refining document processing logic; add chunk validation and OCR capabilities.

+++ src/embedder.py
@@ -3,5 +3,5 @@ from sklearn.feature_extraction.text import TfidfVectorizer
 
 def create_embeddings(chunks):
     vectorizer = TfidfVectorizer()
-    embeddings = vectorizer.fit_transform(chunks).toarray()
+    embeddings = vectorizer.fit_transform(chunks).toarray().astype("float32")
     return embeddings, vectorizer
